Almost final

+++ SMRP_MACAW/SMRP/SMRP.h
@@ -6,7 +6,6 @@
 #include <stdbool.h>
 #include <semaphore.h> // sem_init, sem_wait, sem_trywait, sem_timedwait
 
-
 #include "../MACAW/MACAW.h"
 #include "../common.h"
 #include "../ProtoMon/routing.h"
@@ -20,7 +19,7 @@ typedef struct Routing_Header
     uint8_t src;  // Source of the packet
     uint8_t dst;  // Destination of the packet
     uint8_t prev; // Address of the previous hop
-    int RSSI;     // RSSI of the previous hop address
+    int8_t RSSI;  // RSSI of the previous hop address
 } Routing_Header;
 
 /**
@@ -62,7 +61,6 @@ int SMRP_init(SMRP_Config config);
 int SMRP_sendMsg(uint8_t dest, uint8_t *data, unsigned int len);
 int SMRP_recvMsg(Routing_Header *header, uint8_t *data);
 int SMRP_timedRecvMsg(Routing_Header *header, uint8_t *data, unsigned int timeout);
-
 
 /**
  * @brief Get the address of a random active node from neighbour list.
